Simlified the output and added some logging of the number of audionotes.

+++ Perl/readWavFiles.pl
@@ -38,6 +38,10 @@ print "# by Jeremy Barbay\n\n";
 my $source = "/home/jbarbay/Unison/AudioNotesToProcess/";
 my $destination = "/home/jbarbay/ToArchive/AudioNotesArchived/";
 my $movingFiles=1; # True
+my $debugLevel=0; # 0=silent, 1=print and run all system calls, 2=only print system calls.
+my $logFile="log";
+my $nbFiles;
+my $nbFilesProcessed=0;
 
 if ( @ARGV == 0 ) {
 } elsif ( @ARGV == 1 ) {
@@ -59,137 +63,32 @@ if ( @ARGV == 0 ) {
 
 
 # Estimate of the number of audionotes left to read:
-print("Total number of audio notes to process: ");
-system("ls -l ~/AudioNotesToProcess/*/*.WAV | wc -l");
+$nbFiles = trackNbAudioNotesLeftToRead("$source");
+print "Total number of audio notes to process: $nbFiles \n";
+# system("ls -l $source/*/*.WAV | wc -l");
 
-# print "\n";
-# moveVoiceFolder();  # I kept the code here, but now I use a separate script for this: [[file:~/bin/processDictaphone.pl]]
-
-print "\n";
-# First call to the recursive function:
+# First call to the recursive function which does everything:
 process_dir("");
+# And that's all!
+print "\nThat's all folks!\n";
 
 
+##############################################################################
 
-sub moveVoiceFolder {
-# Automatically Move and rename of "Voice" folder on Dictaphone, and umount at the end
-
-    my $dictaphonePlugged=0; # False
-    my $dictaphoneMount="/media/DICTAPHONE";
-    my $dictaphoneFolder="/media/DICTAPHONE/Record/Voice";
-    
-    if( -e $dictaphoneFolder ) {
-	$dictaphonePlugged=1; # True
-	print("Moving and renaming the file '/media/DICTAPHONE/Record/Voice':\n");
-	
-	# Recover statistics:
-	my $stats = stat($dictaphoneFolder) or die "Error while recovering stats: $!!\n";
-	my $cdate = localtime($stats->ctime);
-	print "Creation Time\t: ",$cdate, "\n";
-	my $mdate = localtime($stats->mtime);
-	print "Modif Time\t: ", $mdate, "\n";
-	my $adate = localtime($stats->atime);
-	print "Access Time\t: ",$adate, "\n";
-	my $backupDate = localtime();
-	print "Back-up Time (today)\t: ", $backupDate, "\n";
-	
-	# Parse supposing $mdate follows the format "Sun Nov 28 06:25:26 2010" 
-	my ($wday, $tmonth, $mday, $hour, $min, $sec, $year) = ($backupDate =~
-								/(\w+)\s+(\w+)\s+(\d+)\s+(\d+):(\d+):(\d+)\s+(\d+)/) or die("Problem parsing modification date: $!!\n");
-	
-	# translate textual month into digital value:
-	my %mon2num = qw(jan 01 feb 02 mar 03 apr 04 may 05 jun 06 jul 07 aug 08 sep 09 oct 10 nov 11 dec 12); 
-	my $month = $mon2num{ lc substr($tmonth, 0, 3) };
-
-	# add a leading zero to the day if less than 10.
-	my $smday = "";
-	if( $mday < 10 ) {
-	    $smday = "0".$mday;
-	} else {
-	    $smday = $mday;
-	}
-
-	# Build new name of Folder:
-	my $baseNewFolderName = $year."-".$month."-".$smday."_".$hour."-".$min."-".$sec."";
-	my $newFolderName = $baseNewFolderName;
-	my $version = 0;
-	while( -e $source.$newFolderName) {
-	    $version = $version+1;
-	    $newFolderName = $baseNewFolderName." version ".$version;
-	}
-	
-	# Move the file to the TOPROCESS folder:
-	print("mv '".$dictaphoneFolder."' '".$source.$newFolderName."'\n");
-	system("mv '".$dictaphoneFolder."' '".$source.$newFolderName."'\n");
-	
-	# Umount the dictaphone
-	print("umount '".$dictaphoneMount."'\n");
-	system("umount '".$dictaphoneMount."'\n");
-    }
-}
-
-
-
-# Action on each wave file
-sub process_file { 
-        my ($f) = shift;                 # file to process
-	my ($relativePath) = shift;      # relative path
-	my ($nbFilesProcessed) = shift;  # rank of the file in the folder
-	my ($nbFiles) = shift;           # total number of wav files in the folder
-	
-
-	if( -e $source.$relativePath.$f ) {
-
-	    # print("CHECKED: file '".$source.$relativePath.$f."' does exit.\n");	
-	    do { 
-		print("Playing file '".$source.$relativePath.$f."'.\n");
-		
-		# Recover and show statistics:
-		my $stats = stat($source.$relativePath.$f) or die "Error while recovering stats: $!!\n";
-		my $cdate = localtime($stats->ctime);
-		print "Creation Time\t: ",$cdate, "\n";
-		my $mdate = localtime($stats->mtime);
-		print "\033[1mLast Modif Time\t: ", $mdate, " \033[0m\n";
-
-		my $filesize = ($stats->size)/1024; # or die "Error while recovering size from stats: $!!\n";	    
-		printf "File Size\t: %uK", $filesize;
-		
-                # Estimate of the number of audionotes left to read:
-		print("\n\n\033[1mNumber of audio notes left: ");
-		system("ls -l ~/AudioNotesToProcess/*/*.WAV | wc -l");
-		print "\033[0m";
-
-		# Play file:
-		system("play  '".$source.$relativePath.$f."'\n");		
-
-		# Take user input.
-		print("Press "
-		      ."'Enter' for repeat,"
-		      ."'Ctrl-d' for next file, "
-		      ."'Ctrl-c' for exit "
-		      ."(".$nbFilesProcessed." processed, "
-		      .($nbFiles-$nbFilesProcessed)." to go in this folder).\n");
-	    } until ( !<STDIN> );	
-	    if ( $movingFiles ) {
-		print("Archiving ".$source.$relativePath.$f." to ".$destination.$relativePath."\n");
-		system("mv '".$source.$relativePath.$f."' '".$destination.$relativePath."'\n");
-	    }
-	} else {
-	    print("WARNING: file '".$source.$relativePath.$f."' does not seem to exit.\n");	    
-	}
-}
-# Recursive Function which does most of the work:
+# Recursive Function which does most of the work
 sub process_dir {
     my ($relativePath) = shift;    # relative path
 
     # find all subdirectories and files
-    print("Opening folder ".$source.$relativePath."\n");
+    if( ($debugLevel == 2) || ($debugLevel == 1) ) {
+	print("Opening folder ".$source.$relativePath."\n");
+    }
     opendir(DIR, $source.$relativePath);
     my @entries = readdir(DIR);
     my @wavFiles = ();
     my @dirs = ();
     my ($e, $f, $d); # iterators for the three loops below
-    my ($nbFiles,$nbFilesProcessed); # Counters
+    my ($nbFilesInThisFolder,$nbFilesProcessedInThisFolder); # Counters
     
     # Sort the content of the folder between folders and WAV files:
     foreach $e (@entries) {
@@ -200,30 +99,41 @@ sub process_dir {
 	    push @wavFiles, $e;
 	}	
     }
-    $nbFiles=@wavFiles;
-    print("Found ".$nbFiles." wav files in folder '".$source.$relativePath."'.\n\n");
+    $nbFilesInThisFolder=@wavFiles;
+    if( ($debugLevel == 2) || ($debugLevel == 1) ) {
+	print("Found ".$nbFilesInThisFolder." wav files in folder '".$source.$relativePath."'.\n\n");
+    }
 
     # Process WAV files:
     my @sfiles = sort { $a cmp $b } @wavFiles;
-    $nbFilesProcessed=0;
+    $nbFilesProcessedInThisFolder=0;
     foreach $f (@sfiles) {
+	process_file($f,$relativePath,$nbFilesProcessedInThisFolder,$nbFilesInThisFolder);
+	$nbFilesProcessedInThisFolder++;
 	$nbFilesProcessed++;
-	process_file($f,$relativePath,$nbFilesProcessed,$nbFiles);
-	print("\n");
+	print "\n";
     }
 
     # Recursively process folders:
     my @sdirs = reverse sort { $a cmp $b } @dirs;
     foreach $d (@sdirs) {
-	print("Process recursively '".$d."'.\n");
+	print("Process recursively '".$d."'.\n\n");
 	if ( $movingFiles ) {
-	    print("mkdir -p ".$destination.$relativePath.$d."\n");
-	    system("mkdir -p '".$destination.$relativePath.$d."'\n");
+	    if( ($debugLevel == 2) || ($debugLevel == 1) ) {
+		print("mkdir -p ".$destination.$relativePath.$d."\n");
+	    }
+	    if( ($debugLevel == 0) || ($debugLevel == 1) ) {
+		system("mkdir -p '".$destination.$relativePath.$d."'\n");
+	    }
 	}
-	process_dir($relativePath.$d."/");
+	process_dir($relativePath.$d."/"); # Recursively process the directory and the files it contains.
 	if ( $movingFiles ) {
-	    print("rmdir ".$source.$relativePath.$d."\n");
-	    system("rmdir '".$source.$relativePath.$d."'\n");
+	    if( ($debugLevel == 2) || ($debugLevel == 1) ) {
+		print("rmdir ".$source.$relativePath.$d."\n");
+	    }
+	    if( ($debugLevel == 0) || ($debugLevel == 1) ) {
+		system("rmdir '".$source.$relativePath.$d."'\n");
+	    }
 	}
     }
 
@@ -231,9 +141,140 @@ sub process_dir {
 
 
 
+# Action on each wave file
+sub process_file { 
+        my ($f) = shift;                 # file to process
+	my ($relativePath) = shift;      # relative path
+	my ($nbFilesProcessedInThisFolder) = shift;  # rank of the file in the folder
+	my ($nbFilesInThisFolder) = shift;           # total number of wav files in the folder
+	
+	if( -e $source.$relativePath.$f ) {
+
+	    do { 
+		# Show how many audionotes processed in this folder.
+		print("\033[1m"
+		      ."Folder: [".($nbFilesProcessedInThisFolder+1)."/".$nbFilesInThisFolder."]; "
+		      ."Session [".($nbFilesProcessed+1)."/".$nbFiles."];"
+		      ."\033[0m\n");
+
+		if( ($debugLevel == 2) || ($debugLevel == 1) ) {
+		    print("Playing file '".$source.$relativePath.$f."'.\n");
+		}
+		
+		# Recover and show statistics:
+		my $stats = stat($source.$relativePath.$f) or die "Error while recovering stats: $!!\n";
+		my $cdate = localtime($stats->ctime);
+		my $mdate = localtime($stats->mtime);
+		my $filesize = ($stats->size)/1024; # or die "Error while recovering size from stats: $!!\n";	    
+		if( ($debugLevel == 2) || ($debugLevel == 1) ) {		
+		    print "Creation Time\t: ",$cdate, "\n";
+		    print "Last Modif Time\t: ", $mdate, "\n";
+		    printf "File Size\t: %uK", $filesize, "\n";
+		}
+
+		# Play file:
+		if( ($debugLevel == 0) ) {
+		    my ($sec,$min,$hour,$mday,$mon,$year,$wday,$yday,$isdst) = localtime($stats->ctime); # Creation time.
+		    print "Note ".$f." taken";
+		    printf(" on %u-%02u-%02u",(1900+$year),($mon+1),$mday);
+		    printf(" at %02u:%02u",$hour,$min);
+		    printf " (of size: %uK)", $filesize;
+		    print ".\n";
+		    system("play -q '".$source.$relativePath.$f."'\n");		
+		} elsif( ($debugLevel == 1) ) {
+		    print "play '".$source.$relativePath.$f."'\n";		
+		    system("play  '".$source.$relativePath.$f."'\n");		
+		} elsif ($debugLevel == 2) {
+		    print "play '".$source.$relativePath.$f."'\n";		
+		}
 
 
+		# Take user input.
+		print("Press "
+		      ."'Enter' for repeat,"
+		      ."'Ctrl-d' for next file, "
+		      ."'Ctrl-c' for exit.\n");
+
+	    } until ( !<STDIN> );	
+	    if ( $movingFiles ) {
+		if( ($debugLevel == 2) || ($debugLevel == 1) ) {
+		    print("Archiving ".$source.$relativePath.$f." to ".$destination.$relativePath."\n");
+		}
+		if( ($debugLevel == 0) || ($debugLevel == 1) ) {
+		    system("mv '".$source.$relativePath.$f."' '".$destination.$relativePath."'\n");
+		}
+	    }
+	} else {
+	    print("WARNING: file '".$source.$relativePath.$f."' does not seem to exit.\n");	    
+	}
+}
 
 
+###################################################################
 
+sub trackNbAudioNotesLeftToRead{
+    my $AudioNotes = shift; # repertory containing the audionotes.
+    my $absoluteTime = time;
+    my ($sec,$min,$hour,$mday,$mon,$year,$wday,$yday,$isdst) = localtime(time);
+    my $count=estimateNbAudioNotesLeftToRead($AudioNotes);
+
+    if( ($debugLevel == 0) | ($debugLevel == 1) ) {
+	if( !(-e "${AudioNotes}${logFile}") ) {
+	    open (LOGFILE, ">${AudioNotes}${logFile}") or die ("Cannot open file ${AudioNotes}${logFile} !!!");
+	    print LOGFILE "# Log produced by the scripts =readWaveFiles.pl= and =processAudioNotes.pl=\n";
+	    print LOGFILE "# ";
+	    print LOGFILE "absoluteTime\t";
+	    print LOGFILE "count\t";
+	    print LOGFILE "yyyy\t";
+	    print LOGFILE "mm-dd\t";
+	    print LOGFILE "hh:mm:ss\t";
+	    print LOGFILE "\n";
+	} else {
+	    open (LOGFILE, ">>${AudioNotes}${logFile}") or die ("Cannot open file ${AudioNotes}${logFile} !!!");
+	}
+	print LOGFILE "$absoluteTime\t";
+	print LOGFILE "$count\t";
+	print LOGFILE sprintf("%u\t",(1900+$year));
+	print LOGFILE sprintf("%02u-%02u\t",($mon+1),$mday);
+	print LOGFILE sprintf("%02u:%02u:%02u\t",$hour,$min,$sec);
+	print LOGFILE "\n";
+	close (LOGFILE);
+    }     
+    if( ($debugLevel == 2) | ($debugLevel == 1) ) {
+	print "open $AudioNotes$logFile\n";
+	print  "absoluteTime\t";
+	print  "count\t";
+	print  "yyyy\t";
+	print  "mm-dd\t";
+	print  "hh:mm:ss\t";
+	print "\n";
+	print  "$absoluteTime\t";
+	print  "$count\t";
+	printf("%u\t",(1900+$year));
+	printf("%02u-%02u\t",($mon+1),$mday);
+	printf("%02u:%02u:%02u\t",$hour,$min,$sec);
+	print "\n";
+	print  'close (LOGFILE);\n';
+    } 
+    return $count;
+}
+
+sub estimateNbAudioNotesLeftToRead {
+    my ($topdir) = shift; # repertory containing the audionotes.
+    opendir (DIR, $topdir) ; 
+    my @entries = readdir(DIR);
+    my @dirs = ();
+    my $totalCount = 0;
+    my $e;
+
+    foreach $e (@entries) {
+	if ( -d "$topdir/$e" && $e ne "." && $e ne ".." ) {
+	    $totalCount = $totalCount + estimateNbAudioNotesLeftToRead("$topdir/$e");
+	}
+	elsif ( -f "$topdir/$e"  && ( ($e =~ /\.wav$/) || ($e =~ /\.WAV$/) )  ) {
+	    $totalCount++;
+	}
+    }
+    return $totalCount;
+}
